requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.0
pandas>=2.0.0
protobuf<5
plotly>=5.18.0
